refactor: Change helper folder location and user defined prop for register navigation when a success register happens
