+++ App.tsx
@@ -1,29 +1,19 @@
 import "react-native-gesture-handler";
 import { NavigationContainer } from "@react-navigation/native";
-import { StyleSheet } from "react-native";
 import { createNativeStackNavigator } from "@react-navigation/native-stack";
-import HomeScreen from "./screens/Home/HomeScreen";
-import RegisterScreen from "./screens/Register/RegisterScreen";
-import type { RootStackParamList } from "./screens/helpers/navigation/NavigationProps";
+import HomeScreen from "./screens/home/HomeScreen";
+import RegisterScreen from "./screens/register/RegisterScreen";
+import type { RootStackParamList } from "./helpers/navigation/NavigationProps";
 
 const Stack = createNativeStackNavigator<RootStackParamList>();
 
 export default function App() {
   return (
     <NavigationContainer>
-      <Stack.Navigator initialRouteName="Home">
+      <Stack.Navigator initialRouteName="Register">
         <Stack.Screen name="Home" component={HomeScreen} />
         <Stack.Screen name="Register" component={RegisterScreen} />
       </Stack.Navigator>
     </NavigationContainer>
   );
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: "#fff",
-    alignItems: "center",
-    justifyContent: "center",
-  },
-});
