+++ helpers/axios/axiosConfig.tsx
@@ -1,0 +1,7 @@
+import axios from "axios";
+
+const instance = axios.create({
+  baseURL: "http://localhost/api",
+});
+
+export default instance;
